Character levels up linked with intelligence

+++ Warlord/Models/ICharacter.cs
@@ -4,11 +4,11 @@ namespace Warlord.Models
     public interface ICharacter
     {
         string Name { get; set; }
-        double Intelligence { get; set; }
-        double Vitality { get; set; }
-        double Strength { get; set; }
-        double Agility { get; set; }
-        double Dominance { get; set; }
+        decimal Intelligence { get; set; }
+        decimal Vitality { get; set; }
+        decimal Strength { get; set; }
+        decimal Agility { get; set; }
+        decimal Dominance { get; set; }
         decimal Experience { get; set; }
         decimal Level { get; set; }
         void GainExp(decimal exp);
